Refactor: move enums all to the same location
