+++ brevitas/core/quant.py
@@ -38,26 +38,20 @@
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 
-from enum import auto
 from typing import Optional, Tuple, Union
 
 import torch
 from torch import Tensor
 from torch.nn import Module
 
-from brevitas.utils.python_utils import AutoName
+from brevitas.inject.enum import QuantType  # retrocompatbility
 from brevitas.function.ops import tensor_clamp, min_int, max_int, max_uint
 from brevitas.function.ops_ste import tensor_clamp_ste, binary_sign_ste, ternary_sign_ste
 
 from .bit_width import BitWidthConst
 from .utils import StatelessBuffer
 
-
-class QuantType(AutoName):
-    BINARY = auto()
-    TERNARY = auto()
-    INT = auto()
-    FP = auto()
+assert QuantType  # prevent removal of unused import
 
 
 class IdentityQuant(torch.jit.ScriptModule):
