+++ brevitas/core/restrict_val.py
@@ -38,7 +38,6 @@
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 
-from enum import auto
 from typing import Callable, Union, Optional
 import math
 
@@ -46,22 +45,12 @@ import torch
 from torch import Tensor
 from torch.nn import Module
 
-from brevitas.utils.python_utils import AutoName
+from brevitas.inject.enum import RestrictValueType, FloatToIntImplType  # retrocompatibility
+
 from .function_wrapper import Identity, PowerOfTwo, LogTwo, ClampMin
 
-
-class RestrictValueType(AutoName):
-    FP = auto()
-    LOG_FP = auto()
-    INT = auto()
-    POWER_OF_TWO = auto()
-
-
-class FloatToIntImplType(AutoName):
-    ROUND = auto()
-    CEIL = auto()
-    FLOOR = auto()
-    ROUND_TO_ZERO = auto()
+assert RestrictValueType  # prevent removal of unused import
+assert FloatToIntImplType
 
 
 class _RestrictClampValue(torch.jit.ScriptModule):
